+++ drivers/pds/lib/utils.go
@@ -1519,23 +1519,20 @@ func CreateDataServiceWorkloads(params WorkloadGenerationParams) (*corev1.Pod, *
 
 	dnsEndpoint, err := GetDeploymentConnectionInfo(params.DeploymentID)
 	if err != nil {
-		log.Errorf("An Error Occured while getting connection info %v", err)
-		return nil, nil, err
+		return nil, nil, fmt.Errorf("error occured while getting connection info %v", err)
 	}
 	log.Infof("Dataservice DNS endpoint %s", dnsEndpoint)
 
 	pdsPassword, err := GetDeploymentCredentials(params.DeploymentID)
 	if err != nil {
-		log.Errorf("An Error Occured while getting credentials info %v", err)
-		return nil, nil, err
+		return nil, nil, fmt.Errorf("error occured while getting credentials info %v ", err)
 	}
 
 	switch params.DataServiceName {
 	case postgresql:
 		dep, err = CreatepostgresqlWorkload(dnsEndpoint, pdsPassword, params.ScaleFactor, params.Iterations, params.DeploymentName, params.Namespace)
 		if err != nil {
-			log.Errorf("An Error Occured while creating postgresql workload %v", err)
-			return nil, nil, err
+			return nil, nil, fmt.Errorf("error occured whilecreating postgresql workload %v", err)
 		}
 
 	case rabbitmq:
@@ -1543,8 +1540,7 @@ func CreateDataServiceWorkloads(params WorkloadGenerationParams) (*corev1.Pod, *
 		command := "while true; do java -jar perf-test.jar --uri amqp://${PDS_USER}:${PDS_PASS}@${AMQP_HOST} -jb -s 10240 -z 100 --variable-rate 100:30 --producers 10 --consumers 50; done"
 		pod, err = CreateRmqWorkload(dnsEndpoint, pdsPassword, params.Namespace, env, command)
 		if err != nil {
-			log.Errorf("An Error Occured while creating rabbitmq workload %v", err)
-			return nil, nil, err
+			return nil, nil, fmt.Errorf("error occured while creating rabbitmq workload %v ", err)
 		}
 
 	case redis:
@@ -1552,23 +1548,22 @@ func CreateDataServiceWorkloads(params WorkloadGenerationParams) (*corev1.Pod, *
 		command := "redis-benchmark -a ${PDS_PASS} -h ${REDIS_HOST} -r 10000 -c 1000 -l -q --cluster --user ${PDS_USER}"
 		pod, err = CreateRedisWorkload(params.DeploymentName, redisStressImage, dnsEndpoint, pdsPassword, params.Namespace, env, command)
 		if err != nil {
-			log.Errorf("An Error Occured while creating redis workload %v", err)
-			return nil, nil, err
+			return nil, nil, fmt.Errorf("error occured while creating redis workload %v ", err)
 		}
 
 	case cassandra:
-		cassCommand := params.DeploymentName + " write no-warmup n=1000000 cl=ONE -mode user=pds password=" + pdsPassword + " native cql3 -col n=FIXED\\(5\\) size=FIXED\\(64\\)  -pop seq=1..1000000 -node " + dnsEndpoint + " -port native=9042 -rate auto -log file=/tmp/" + params.DeploymentName + ".load.data -schema \"replication(factor=3)\" -errors ignore; cat /tmp/" + params.DeploymentName + ".load.data"
+		cassCommand := fmt.Sprintf("%s write no-warmup n=1000000 cl=ONE -mode user=pds password=%s native cql3 -col n=FIXED\\(5\\) size=FIXED\\(64\\)  -pop seq=1..1000000 -node %s -port native=9042 -rate auto -log file=/tmp/%s.load.data -schema \"replication(factor=3)\" -errors ignore; cat /tmp/%s.load.data", params.DeploymentName, pdsPassword, dnsEndpoint, params.DeploymentName, params.DeploymentName)
+		//cassCommand := params.DeploymentName + " write no-warmup n=1000000 cl=ONE -mode user=pds password=" + pdsPassword + " native cql3 -col n=FIXED\\(5\\) size=FIXED\\(64\\)  -pop seq=1..1000000 -node " + dnsEndpoint + " -port native=9042 -rate auto -log file=/tmp/" + params.DeploymentName + ".load.data -schema \"replication(factor=3)\" -errors ignore; cat /tmp/" + params.DeploymentName + ".load.data"
 		dep, err = CreateDeploymentWorkloads(cassCommand, params.DeploymentName, cassandraStresImage, params.Namespace)
 		if err != nil {
-			log.Errorf("An Error Occured while creating cassandra workload %v", err)
-			return nil, nil, err
+			return nil, nil, fmt.Errorf("error occured while creating cassandra workload %v ", err)
 		}
 	case elasticSearch:
-		esCommand := "while true; do esrally race --track=geonames --target-hosts=" + dnsEndpoint + " --pipeline=benchmark-only --test-mode --kill-running-processes --client-options=\"timeout:" + params.TimeOut + ",use_ssl:" + params.UseSSL + ",verify_certs:" + params.VerifyCerts + ",basic_auth_user:" + params.User + ",basic_auth_password:" + pdsPassword + "\"; done"
+		esCommand := fmt.Sprintf("while true; do esrally race --track=geonames --target-hosts=%s --pipeline=benchmark-only --test-mode --kill-running-processes --client-options=\"timeout:%s,use_ssl:%s,verify_certs:%s,basic_auth_user:%s,basic_auth_password:%s; done", dnsEndpoint, params.TimeOut, params.UseSSL, params.VerifyCerts, params.User, pdsPassword)
+		//esCommand := "while true; do esrally race --track=geonames --target-hosts=" + dnsEndpoint + " --pipeline=benchmark-only --test-mode --kill-running-processes --client-options=\"timeout:" + params.TimeOut + ",use_ssl:" + params.UseSSL + ",verify_certs:" + params.VerifyCerts + ",basic_auth_user:" + params.User + ",basic_auth_password:" + pdsPassword + "\"; done"
 		dep, err = CreateDeploymentWorkloads(esCommand, params.DeploymentName, esRallyImage, params.Namespace)
 		if err != nil {
-			log.Errorf("An Error Occured while creating elasticSearch workload %v", err)
-			return nil, nil, err
+			return nil, nil, fmt.Errorf("error occured while creating elasticSearch workload %v ", err)
 		}
 
 	case couchbase:
@@ -1580,8 +1575,7 @@ func CreateDataServiceWorkloads(params WorkloadGenerationParams) (*corev1.Pod, *
 
 		pod, err = CreatePodWorkloads(params.DeploymentName, cbloadImage, params, params.Namespace, "1000", env)
 		if err != nil {
-			log.Errorf("An Error Occured while creating couchbase workload %v", err)
-			return nil, nil, err
+			return nil, nil, fmt.Errorf("error occured while creating couchbase workload %v ", err)
 		}
 	}
 	return pod, dep, nil
