+++ tests/pds/common.go
@@ -1,0 +1,131 @@
+package tests
+
+import (
+	pds "github.com/portworx/pds-api-go-client/pds/v1alpha1"
+	"github.com/portworx/sched-ops/k8s/core"
+	pdslib "github.com/portworx/torpedo/drivers/pds/lib"
+	"github.com/portworx/torpedo/pkg/aetosutil"
+	"github.com/portworx/torpedo/pkg/log"
+	v1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+)
+
+type PDSDataService struct {
+	Name          string "json:\"Name\""
+	Version       string "json:\"Version\""
+	Image         string "json:\"Image\""
+	Replicas      int    "json:\"Replicas\""
+	ScaleReplicas int    "json:\"ScaleReplicas\""
+	OldVersion    string "json:\"OldVersion\""
+	OldImage      string "json:\"OldImage\""
+}
+
+const (
+	pdsNamespace            = "pds-system"
+	deploymentName          = "qa"
+	envDeployAllDataService = "DEPLOY_ALL_DATASERVICE"
+	postgresql              = "PostgreSQL"
+	cassandra               = "Cassandra"
+	elasticSearch           = "Elasticsearch"
+	couchbase               = "Couchbase"
+	redis                   = "Redis"
+	rabbitmq                = "RabbitMQ"
+	mongodb                 = "MongoDB"
+	mysql                   = "MySQL"
+	kafka                   = "Kafka"
+	zookeeper               = "ZooKeeper"
+	pdsNamespaceLabel       = "pds.portworx.com/available"
+)
+
+var (
+	namespace                               string
+	pxnamespace                             string
+	tenantID                                string
+	dnsZone                                 string
+	projectID                               string
+	serviceType                             string
+	deploymentTargetID                      string
+	replicas                                int32
+	err                                     error
+	supportedDataServices                   []string
+	dataServiceNameDefaultAppConfigMap      map[string]string
+	namespaceID                             string
+	storageTemplateID                       string
+	dataServiceDefaultResourceTemplateIDMap map[string]string
+	dataServiceNameIDMap                    map[string]string
+	supportedDataServicesNameIDMap          map[string]string
+	DeployAllVersions                       bool
+	DataService                             string
+	DeployAllImages                         bool
+	dataServiceDefaultResourceTemplateID    string
+	dataServiceDefaultAppConfigID           string
+	dataServiceVersionBuildMap              map[string][]string
+	dataServiceImageMap                     map[string][]string
+	pdsPods                                 []corev1.Pod
+	dep                                     *v1.Deployment
+	pod                                     *corev1.Pod
+	params                                  *pdslib.Parameter
+	podList                                 *corev1.PodList
+	isDeploymentsDeleted                    bool
+	isNamespacesDeleted                     bool
+	dash                                    *aetosutil.Dashboard
+	deployment                              *pds.ModelsDeployment
+	k8sCore                                 = core.Instance()
+	pdsLabels                               = make(map[string]string)
+)
+
+var dataServiceDeploymentWorkloads = []string{cassandra, elasticSearch, postgresql}
+var dataServicePodWorkloads = []string{redis, rabbitmq, couchbase}
+
+func RunWorkloads(params pdslib.WorkloadGenerationParams, ds PDSDataService, deployment *pds.ModelsDeployment, namespace string) (*corev1.Pod, *v1.Deployment, error) {
+	params.DataServiceName = ds.Name
+	params.DeploymentID = deployment.GetId()
+	params.Namespace = namespace
+
+	if ds.Name == postgresql {
+		params.DeploymentName = "pgload"
+		params.ScaleFactor = "100"
+		params.Iterations = "1"
+
+		log.Infof("Running Workloads on DataService %v ", ds.Name)
+		pod, dep, err = pdslib.CreateDataServiceWorkloads(params)
+
+	}
+	if ds.Name == rabbitmq {
+		params.DeploymentName = "rmq"
+		log.Infof("Running Workloads on DataService %v ", ds.Name)
+		pod, dep, err = pdslib.CreateDataServiceWorkloads(params)
+
+	}
+	if ds.Name == redis {
+		params.DeploymentName = "redisbench"
+		log.Infof("Running Workloads on DataService %v ", ds.Name)
+		pod, dep, err = pdslib.CreateDataServiceWorkloads(params)
+
+	}
+	if ds.Name == cassandra {
+		params.DeploymentName = "cassandra-stress"
+		log.Infof("Running Workloads on DataService %v ", ds.Name)
+		pod, dep, err = pdslib.CreateDataServiceWorkloads(params)
+
+	}
+	if ds.Name == elasticSearch {
+		params.DeploymentName = "es-rally"
+		params.User = "elastic"
+		params.UseSSL = "false"
+		params.VerifyCerts = "false"
+		params.TimeOut = "60"
+		log.Infof("Running Workloads on DataService %v ", ds.Name)
+		pod, dep, err = pdslib.CreateDataServiceWorkloads(params)
+
+	}
+	if ds.Name == couchbase {
+		params.DeploymentName = "cb-load"
+		log.Infof("Running Workloads on DataService %v ", ds.Name)
+		pod, dep, err = pdslib.CreateDataServiceWorkloads(params)
+
+	}
+
+	return pod, dep, err
+
+}
